requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
sqlparse==0.5.1
typing-extensions==4.12.2
psycopg2-binary==2.9.2
python-decouple==3.3
dj-database-url==2.2.0
requests==2.32.3
djangorestframework==3.15.2